Skill agregado y en funcionamiento

+++ src/main/java/com/porfolio/PRM/Entity/EEducacion.java
@@ -6,8 +6,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import lombok.Getter;
-import lombok.Setter;
 
 
 @Entity
@@ -15,7 +13,7 @@ public class EEducacion {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     
     @NotNull
     @Size (min = 1, max = 75, message="El nombre no cumple con la longitud.")
@@ -38,11 +36,11 @@ public class EEducacion {
         this.periodo = periodo;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
